+++ FlagsGen7USUM.cs
@@ -29,7 +29,7 @@ namespace MissingEventFlagsCheckerPlugin
                 if (idx >= 0x9D8 && idx < 0xBB0)
                 {
                     int i = idx - 0x9D8;
-                    m_eventFlagsList.Add(new FlagDetail(idx, FlagType.FieldItem, "", i.ToString("D3") + $" (Flag 0x{(idx).ToString("X3")})") { IsSet = savEventFlags[idx] });
+                    m_eventFlagsList.Add(new FlagDetail((uint)idx, FlagType.FieldItem, "", i.ToString("D3") + $" (Flag 0x{(idx).ToString("X3")})") { IsSet = savEventFlags[idx] });
                 }
 
                 // FlagType.TrainerBattle:
@@ -40,13 +40,13 @@ namespace MissingEventFlagsCheckerPlugin
                 else if (idx >= 0xC00 && idx < 0xE70)
                 {
                     int i = idx - 0xC00;
-                    m_eventFlagsList.Add(new FlagDetail(idx, FlagType.TrainerBattle, "", i.ToString("D3") + $" (Flag 0x{(idx).ToString("X3")})") { IsSet = savEventFlags[idx] });
+                    m_eventFlagsList.Add(new FlagDetail((uint)idx, FlagType.TrainerBattle, "", i.ToString("D3") + $" (Flag 0x{(idx).ToString("X3")})") { IsSet = savEventFlags[idx] });
                 }
 
                 // Misc
                 else
                 {
-                    m_eventFlagsList.Add(new FlagDetail(idx, FlagType._Unknown, "", "") { IsSet = savEventFlags[idx] });
+                    m_eventFlagsList.Add(new FlagDetail((uint)idx, FlagType._Unknown, "", "") { IsSet = savEventFlags[idx] });
                 }
             }
 
@@ -87,7 +87,7 @@ namespace MissingEventFlagsCheckerPlugin
                     if (f.FlagTypeVal == flagType)
                     {
                         f.IsSet = value;
-                        flagHelper.SetEventFlag(f.FlagIdx, value);
+                        flagHelper.SetEventFlag((int)f.FlagIdx, value);
                     }
                 }
             }
